Implement 'can_substitute' metafunction.
Introduces a hack-y "SuppressDiagnostics" mechanism to "Sema" which,
when set (via RAII class) suppresses all diagnostics. Without this, the
'CheckTemplateArgumentList' machinery will emit diagnostics while
evaluating 'can_substitute' for an invalid template-id. After first
trying to introduce a 'Complain'-parameter that suppresses diagnostics
when false, and thereby coming to understand what an invasive change
that would be, this felt like a better solution.