Refactor value/object reflection and CXXReflectExpr.
Store reflections of values and objects as pointers to APValues instead
of as ConstantExprs. In general, try to stop abusing the poor
ConstantExpr. In the process, I also refactored the CXXReflectExpr and
some of the reflection API surface of APValue. A little more explicit
handling of reflection values, but I think a good overall result.

I had previously added small patches in many places to check for an
"empty" ConstantExpr (i.e., one without a subexpression) which diverge
from upstream. I think most of these checks can now be removed; finding
them is another matter.

There's certainly room for more cleanup. It would be great if the
APValues for reflections of values/objects could be transient (rather
than persisted in AST memory), but the mutually recursive nesting of
APValues and ReflectionValues within one another makes this tricky.
Caching them in the ASTContext (a la TypeInfos and the like) is one
idea, which wouldn't require much effort. Anyway, this is still an
improvement over the prior abused ConstantExpr approach.

Won't be surprised if there are some bugs to shake out with the large
reworking of things, but at least the tests pass.