Return a proper character array for 'define_static_string'.
This is somewhat tricky, since the character array needs to be
associated to a declaration (i.e., a hidden variable), but the
declaration won't be parsed and therefore won't get handed off to
CodeGen. So unless we do something to tell CodeGen about the variable,
it won't get emitted and we'll get linker errors.

So we (for now, anyway) introduce a similar mechanism to what's already
used for "#pragma weak"; we track any generated decl's in an array held
by 'Sema', and pass those explicitly to CodeGen after parsing the AST.

We also cache any such generated variables in the ASTContext, partly as
an optimization, but mostly to ensure that each such variable obtains a
(TU-)unique name.

Closes issue #80.