Refactor of reflection representation.
Merge 'ReflectionValue' into 'APValue' entirely. Add a "reflection
depth" counter to every APValue, such that if APValues V and RV are the
same except that RV has a ReflectionDepth one higher than V, then RV
represents a reflection of V.

This permits the representation of reflections of values and objects
without dynamically allocating any memory to store the reflected
results.